update api medical record

+++ Modules/Pemedic/Services/UserService.php
@@ -58,12 +58,19 @@ class UserService {
         $oldUser =  $this->userRepository->findByAttributes(['email'=>$request->email]);
         if(!empty($oldUser))
         {
-            $oldUser->roles()->attach($role);
+            if($oldUser->roles()->first()->id != config('asgard.userprofile.config.roles.clinic'))
+            {
+                $oldUser->roles()->attach($role);
+            }
+            else
+            {
+                return null;
+            }
         }
         else
         {
             $user = $this->userRepository->addNewUser($role,$data,0);
-            if(count($data['clinic_id'])>0)
+            if(isset($data['clinic_id']) && count($data['clinic_id'])>0)
             {
                  // assign clinic
                 $user->clinic()->attach($data['clinic_id']);
@@ -83,7 +90,7 @@ class UserService {
     {
         $data['dob'] = !empty($data['dob'])?date('Y-m-d', strtotime($data['dob'])):null;
         $this->userRepository->update($user,$data);
-        if(count($data['clinic_id'])>0)
+        if(isset($data['clinic_id']) && count($data['clinic_id'])>0)
         {
             $user->clinic()->detach();
             $user->clinic()->attach($data['clinic_id']);
@@ -108,6 +115,72 @@ class UserService {
               $sheet->loadView('pemedic::admin.csvs.export-csv',['users' => $users]);
             });
         })->download('csv');
+    }
+
+    /**
+     * import patients csv function
+     * @param request
+     */
+    public function import($request)
+    {   
+        if($request->hasFile('file')) {
+            $items = \Excel::load($request->file('file'), function($reader) {
+            })->get()->values()->toArray();
+            if(count($items)>0)
+            {
+                $dataErrors = [];
+                $dataArr = [];
+                $totalSuccess = 0;
+                $formatErrors =[];
+                // First Validate CSV, if failed, prevent csv upload.
+                foreach($items as $key=>$value) {
+                    $userCsvDetail = array_values($value); 
+                    $validateCsvResult = $this->validateCsvFormat($userCsvDetail);
+                    if($validateCsvResult != "Valid") {
+                        $formatErrors[] = "Row ".($key + 2)." : ".$validateCsvResult;
+                    }
+                    $validate_email = $this->userRepository->validateEmail(config('asgard.userprofile.config.roles.patient'),$userCsvDetail[0]);
+                    if($validate_email)
+                    {
+                        $validate_phone = $this->userProfileRepository->findByAttributes(['phone'=>$userCsvDetail[1]]);
+                        if(empty($validate_phone))
+                        {
+                            $data = [];
+                            $oldUser =  $this->userRepository->findByAttributes(['email'=>$userCsvDetail[0]]);
+                            if(!empty($oldUser))
+                            {
+                                $totalSuccess++;
+                                // $oldUser->roles()->attach($role);
+                            }
+                            else
+                            {
+                                $totalSuccess++;
+                                $password = $this->generatePassword();
+                                $data['email']      = $userCsvDetail[0];
+                                $data['password']   = Hash::make($password);
+                                $data['phone']      = $userCsvDetail[1];
+                                // $user = $this->userRepository->addNewUser(config('asgard.userprofile.config.roles.patient'),$data,0);
+                            }
+                        }
+                        else
+                        {
+                            $formatErrors[] = "Row ".($key + 2)." : The phone has already been taken";
+                            $dataErrors[] = $userCsvDetail[0];
+                        }
+                    }
+                    else
+                    {
+                        $formatErrors[] = "Row ".($key + 2)." : The email has already been taken";
+                        $dataErrors[] = $userCsvDetail[0];
+                    }                
+                }
+                $dataArr['totalSuccess']    = $totalSuccess;
+                $dataArr['dataErrors']      = $dataErrors;
+                $dataArr['formatErrors']    = $formatErrors;
+                return $dataArr;
+            }
+        }
+        
     }
 
     /**
@@ -147,6 +220,12 @@ class UserService {
             $user = $this->userProfileRepository->find($id);
             if($user)
             {
+                $user->phone = $user->phone . time();
+                $user->save();
+
+                $patient = $user->user;
+                $patient->email = $patient->email . time();
+                $patient->save();
                 $this->userProfileRepository->destroy($user);
                 $this->userRepository->destroy($user->user);
             }
@@ -169,5 +248,32 @@ class UserService {
         {
             return $url;
         } 
+    }
+
+    /**
+     * validate field email
+     * @param $dataUser
+     * @return message/null
+     */
+    private function validateCsvFormat($dataUser) {
+        // Invalidate data if missing columns
+        if(sizeof($dataUser) != 2) {
+            return "Import failed due to the wrong number of columns";
+        }
+
+        // Regex
+        $invalidFields = [];
+
+        if(!filter_var($dataUser[0], FILTER_VALIDATE_EMAIL)){
+            array_push($invalidFields, 'Email'); 
+        }
+        if(!preg_match('/^[0-9]*$/', $dataUser[1])){ array_push($invalidFields, 'Phone'); }
+        
+        // Return true if no errors otherwise return error message
+        if(count($invalidFields) == 0) {
+            return "Valid";
+        } else {
+            return $dataUser[0]." import failed due to the following invalid fields: ".implode(", ", $invalidFields);
+        }
     }
 }
